rework and extend arkade script e2e tests
